Added a new way to add things to the zetteldesk + some minor changes
Full list of changes: Added the
zetteldesk-add-moc-or-poi-backlink-to-desktop function (look at the
docstring for what it does) and made zetteldesk-add-node-to-desktop
have an input argument and its org-roam-node-read part as a part of
the interactive call. This is to be used with the above function to
avoid too much needless copy-pasting. Also defined and alternative
org-roam-backlink-query function which takes a node and doesn't just
use the node at point.