Finally managed to integrate org-remark to the zetteldesk workflow.
The code for that is in zetteldesk-remark.el. Its definitely in its
infancy still, but it looks like a working product so I am really
happy with it. I plan to write the documentation for it in the wiki
when I have time (hopefully soon). Also I made some small changes to
the README of the repo in this commit.