Minor fixes to zetteldesk-remark and improvements to zetteldesk-ref
In remark integration I added a function to reset the notes file to
its original value in the off hook of the zetteldesk-remark-mode, and
another to ensure the heading it takes is a top level one. In
zeteldesk-ref I added filter functions with helm and ivy UIs instead
of only the org-roam one.