Fixing an outdated section of the README as per #8.
Specifically, the doom installation instructions were really old (as
someone had written them for me like a year ago) and not true anymore
because at that point the package wasn't on MELPA so adding the
package to straight meant you would install all the extensions as
well. How it is now, means that each package is its own MELPA recipe
so that instruction caused a broken config. I have corrected it now to
mention that each one is a separate package and referred to #8 for any
config problems.